+++ src/InvisibleParam/CS/Command.cs
@@ -24,23 +24,22 @@
 using System;
 using System.IO;
 using System.Reflection;
+using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
-
 
 namespace Revit.SDK.Samples.InvisibleParam.CS
 {
     /// <summary>
-    /// This command show how to use parameter file to store invisible and 
-    /// visible parameters.
+    ///     This command show how to use parameter file to store invisible and
+    ///     visible parameters.
     /// </summary>
-    [Autodesk.Revit.Attributes.Transaction(Autodesk.Revit.Attributes.TransactionMode.Manual)]
-    [Autodesk.Revit.Attributes.Regeneration(Autodesk.Revit.Attributes.RegenerationOption.Manual)]
-    [Autodesk.Revit.Attributes.Journaling(Autodesk.Revit.Attributes.JournalingMode.NoCommandData)]
+    [Transaction(TransactionMode.Manual)]
+    [Regeneration(RegenerationOption.Manual)]
+    [Journaling(JournalingMode.NoCommandData)]
     public class Command : IExternalCommand
     {
-
-                public Result Execute(
+        public Result Execute(
             ExternalCommandData commandData, ref string message, ElementSet elements)
         {
             var transaction = new Transaction(commandData.Application.ActiveUIDocument.Document, "External Tool");
@@ -51,14 +50,11 @@ namespace Revit.SDK.Samples.InvisibleParam.CS
                 //Create a clear file as parameter file.
                 var path = Assembly.GetExecutingAssembly().Location;
                 var index = path.LastIndexOf("\\");
-                var newPath = path.Substring(0,index);
+                var newPath = path.Substring(0, index);
                 newPath += "\\RevitParameters.txt";
-                if(File.Exists(newPath))
-                {
-                    File.Delete(newPath);
-                }
+                if (File.Exists(newPath)) File.Delete(newPath);
                 var fs = File.Create(newPath);
-                fs.Close();               
+                fs.Close();
 
                 //cache application handle
                 var revitApp = commandData.Application.Application;
@@ -69,7 +65,9 @@ namespace Revit.SDK.Samples.InvisibleParam.CS
                 var parafile = revitApp.OpenSharedParameterFile();
 
                 //get walls category
-                var wallCat = commandData.Application.ActiveUIDocument.Document.Settings.Categories.get_Item(BuiltInCategory.OST_Walls);
+                var wallCat =
+                    commandData.Application.ActiveUIDocument.Document.Settings.Categories.get_Item(BuiltInCategory
+                        .OST_Walls);
                 var categories = revitApp.Create.NewCategorySet();
                 categories.Insert(wallCat);
 
@@ -79,20 +77,22 @@ namespace Revit.SDK.Samples.InvisibleParam.CS
                 var apiGroup = parafile.Groups.Create("APIGroup");
 
                 //Create a visible "VisibleParam" of text type.
-                var ExternalDefinitionCreationOptions1 = new ExternalDefinitionCreationOptions("VisibleParam", SpecTypeId.String.Text);
+                var ExternalDefinitionCreationOptions1 =
+                    new ExternalDefinitionCreationOptions("VisibleParam", SpecTypeId.String.Text);
                 var visibleParamDef = apiGroup.Definitions.Create
                     (ExternalDefinitionCreationOptions1);
-                                        ;
-                                        var bindingMap = commandData.Application.ActiveUIDocument.Document.ParameterBindings;
+                ;
+                var bindingMap = commandData.Application.ActiveUIDocument.Document.ParameterBindings;
                 bindingMap.Insert(visibleParamDef, binding);
 
                 //Create a invisible "InvisibleParam" of text type.
-                var ExternalDefinitionCreationOptions2 = new ExternalDefinitionCreationOptions("InvisibleParam", SpecTypeId.String.Text);
+                var ExternalDefinitionCreationOptions2 =
+                    new ExternalDefinitionCreationOptions("InvisibleParam", SpecTypeId.String.Text);
                 var invisibleParamDef = apiGroup.Definitions.Create
                     (ExternalDefinitionCreationOptions2);
                 bindingMap.Insert(invisibleParamDef, binding);
             }
-            catch(Exception e)
+            catch (Exception e)
             {
                 transaction.RollBack();
                 message = e.ToString();
@@ -102,8 +102,8 @@ namespace Revit.SDK.Samples.InvisibleParam.CS
             {
                 transaction.Commit();
             }
+
             return Result.Succeeded;
         }
-
-            }
+    }
 }
